Fix phone account isn't associated
While trying to start a MT/MO call, the phone accounts might be
in process of tearing down or setup up, so that the call is unable
to continue.

To fix it, always associate MT notifier with Phone, which is not
required to be created along with account entry, besides, check if
accounts need to be tore down to reduce the occurrence rate of being
in no account found.

Change-Id: I4a3409cb3c63babe2780ea888eaa19985a97c0ce
CRs-Fixed: 2813916