IMS: Do not show wrong prompt in case of VT upgrade request.
When an incoming call is received on the other sub, prompt
"Answering this call will end your other call" is shown.
Once the call on other sub is active and an incoming modify
request is received for the call, the same prompt is wrongly
shown again.

Fix this by removing the extra once the incoming call transitions
from INCOMING state to some other state.

Change-Id: I3f2c4b566fa6f499b508129b92c7d5101b8ac0e6
CRs-Fixed: 2820595