Query call forward status properly
Currently QtiImsExtManager is used to query the call forward status
only once from startCallForwardOptionsQuery due to the check for the
mQtiImsExtConnector null or not. If the startCallForwardOptionsQuery
is used to query again (for example in case of CFNRc changed, CFNL
also need to be queried again) it uses ExtTelephonyManager/Phone APIs
instead of using QtiImsExtManager.
Instead of this always use QtiImsExtManager to query call forward
status if UT is enabled so that IMS specific response is properly received.

Change-Id: Ifbcdeafe8cd6559617ea233ccb6986d8ef5565d5
CRs-Fixed: 3092521