Fix voice call preference changing to different slot
On a Subsidy lock enabled device, it is seen that till
the subsidy carrier sim latches to network, the other
sim stays in perso locked state, due to this subsidy
carrier sim always initialized first when compared to
non-subsidy sim. This is resulting in voice preference
getting set to subsidy carrier sim.

Change-Id: I6c387bc7309e149d77cc564a557eadf247e84938
CRs-Fixed: 3140714